# openapi/openapi.yaml
openapi: '3.1.0'
info:
  title: Narrex API
  version: '0.1.0'
  description: |
    API for Narrex — a visual novel editor with AI scene generation.

    Phase 1 covers the core loop: project creation with auto-structuring,
    multi-track timeline, character map, and scene-level AI draft generation.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/auth/google:
    get:
      operationId: initiateGoogleAuth
      tags: [auth]
      summary: Redirect to Google OAuth consent screen
      security: []
      parameters:
        - name: redirect_uri
          in: query
          required: false
          description: URL to redirect after authentication completes
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to Google OAuth
          headers:
            Location:
              schema:
                type: string
              description: Google OAuth consent URL
  /v1/auth/google/callback:
    get:
      operationId: handleGoogleCallback
      tags: [auth]
      summary: Handle Google OAuth callback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to web app with tokens set
          headers:
            Location:
              schema:
                type: string
              description: Web app URL
            Set-Cookie:
              schema:
                type: string
              description: httpOnly refresh token cookie (30 days)
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      description: Uses httpOnly refresh token cookie. No request body needed.
      security: []
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfile'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Delete current user's account and all data
      description: |
        Permanently deletes the authenticated user's account and all associated
        data (projects, scenes, characters, drafts, generation logs).
        Clears the refresh token cookie. Irreversible.
      responses:
        '204':
          description: Account deleted
          headers:
            Set-Cookie:
              description: Clears the refresh token cookie
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/auth/me/avatar:
    post:
      operationId: uploadAvatar
      tags: [auth]
      summary: Upload profile avatar image
      description: |
        Accepts a multipart file upload (field name: `avatar`).
        Supported formats: JPEG, PNG, WebP. Max size: 2MB.
        The image is stored in R2 and the user's profileImageUrl is updated.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, or WebP, max 2MB)
      responses:
        '200':
          description: Updated user with new avatar URL
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Self-issued JWT access token (15 min lifetime)

  # ---- Reusable Parameters ------------------------------------------------
...
  responses:
    BadRequest:
      description: Malformed request syntax
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: 'https://api.narrex.app/errors/validation-failed'
        title:
          type: string
          example: Validation Failed
        status:
          type: integer
          example: 422
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        requestId:
          type: string
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
          example: title
        code:
          type: string
          example: required
        message:
          type: string
          example: Title is required
...
    User:
      type: object
      required: [id, email, displayName, themePreference, languagePreference, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        displayName:
          type: string
        profileImageUrl:
          type:
            - string
            - 'null'
          format: uri
        themePreference:
          type: string
          enum: [system, light, dark]
          default: system
          description: User's theme preference
        languagePreference:
          type: string
          enum: [ko, en]
          default: ko
          description: User's language preference
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    UpdateProfile:
      type: object
      description: Profile image URL is only settable via the avatar upload endpoint.
      properties:
        displayName:
          type:
            - string
            - 'null'
        themePreference:
          type: string
          enum: [system, light, dark]
        languagePreference:
          type: string
          enum: [ko, en]
    AuthTokens:
      type: object
      required: [accessToken, expiresIn]
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          description: Token lifetime in seconds
          example: 900

    # ---- Project ----------------------------------------------------------
...
security:
  - bearerAuth: []